requirements.txt
names~=0.3.0
pdfrw~=0.4
PyYAML~=5.4.1